Backlog #11 är klar + test

+++ Service/src/main/java/sportstats/service/util/GameByTeam.java
@@ -50,7 +50,7 @@ public class GameByTeam {
         return delegate.getDate();
     }
       
-    public Long getSpectators() {
+    public int getSpectators() {
         return delegate.getSpectators();
     }
